move arm template deploy to util
